replace integer with long
co-276

+++ src/main/groovy/edu/oregonstate/mist/directoryapi/DirectoryEntityDAO.groovy
@@ -45,7 +45,7 @@ class DirectoryEntityDAO {
     /**
      * Returns directory entity matching input id.
      */
-    public DirectoryEntity getByOSUUID(Integer osuuid) {
+    public DirectoryEntity getByOSUUID(Long osuuid) {
         // TODO: convert parameter into filter string
         // TODO: return search result
     }
